requirements.txt
asgiref==3.7.2
aspose-words==23.11.0
asttokens==2.4.0
backcall==0.2.0
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
crypto==1.4.1
cryptography==41.0.4
cssbeautifier==1.14.9
cycler==0.12.1
Cython==3.0.0
debugpy==1.8.0
decorator==5.1.1
Django==5.0.1
djlint==1.34.0
EditorConfig==0.12.3
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
fire==0.5.0
fonttools==4.43.1
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
IMAPClient==2.3.1
imbox==0.9.8
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
joblib==1.3.2
jsbeautifier==1.14.9
json5==0.9.14
jupyter_client==8.4.0
jupyter_core==5.4.0
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
Naked==0.1.32
nest-asyncio==1.5.8
numpy==1.26.0
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
parso==0.8.3
pathspec==0.11.2
pdf2docx==0.5.6
pdf2image==1.17.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyMuPDF==1.23.4
PyMuPDFb==1.23.3
pyodbc==4.0.39
pyparsing==3.1.1
PyPDF2==3.0.1
PyPDF4==1.27.0
pyrfc==3.0
pytesseract==0.3.9
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
reportlab==4.0.5
requests==2.31.0
schedule==1.2.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
termcolor==2.3.0
threadpoolctl==3.2.0
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.5
wcwidth==0.2.8
xmltodict==0.13.0
